backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-genai
chromadb
python-dotenv
pandas
pydantic
watchdog
